sha1: Change prototype of sha1_file to match implementation

+++ sha1.h
@@ -10,7 +10,7 @@
 #define SHA1_Update	blk_SHA1_Update
 #define SHA1_Final	blk_SHA1_Final
 
-int sha1_file(FILE *infile, unsigned long size, unsigned char *sha1);
+int sha1_file(FILE *infile, size_t size, unsigned char *sha1);
 void print_sha1(char *dst, unsigned char *sha1);
 
 #endif
